Moved core to core.py
__version__ is now in __init__.py (as is should be) and the core is
defined in core.py. Other than version, __init__.py is just a dummy
which imports the pieces we make public.
